feat: move optimization_params under backtesting module
