+++ src/strategy_tester/backtesting/optimization.py
@@ -2,7 +2,7 @@ from typing import Type, Tuple
 import pandas as pd
 from backtesting import Backtest, Strategy
 
-from ..optimization_params import OptimizationParams
+from .optimization_params import OptimizationParams
 from .broker_params import BrokerParams
 
 def run_optimization(
