+++ src/strategy_tester/backtesting/optimization_params.py
